LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-anchor-multicol-display.tentative-expected.txt


FAIL .target 1 assert_equals:
<div class="target" popover="" anchor="anchor1" data-expected-height="50"></div>
height expected 50 but got 20
